Simplify: remove forms-data-cache wrapper, inline invalidation calls
Delete forms-data-cache.ts and move getFormsListQuery back into
use-forms-data.ts. Inline invalidateResolutionForStoreSelector
calls directly at each call site instead of routing through a
single-line wrapper function.
